Added instance-level callbacks for ActiveRecord models.
Changed pusher and view nodes so that it is not the pusher who is
listening for all model changes, but it is the view nodes each
listening only for relevant (at the individual record level) changes
and then reporting back to pusher when a change happens.

Replaced record renderer controller with an equivalent event.

Made RedAppController clear client autoviews before every request,
unless the current controller is "async".